AES template class keylength measured in bytes

+++ src/aes.cpp
@@ -212,7 +212,7 @@ template <uint8 k_len> void AES<k_len>::update(uint8 in[16]) {
 /* Generate the key schedule for each state transformation round
 Parameter: uint8 key[32]: reference to the key inputed by the user
 */
-template <uint32 k_len> void AES<k_len>::key_expansion(uint8 key[k_len]) {
+template <uint8 k_len> void AES<k_len>::key_expansion(uint8 key[k_len]) {
   uint8 Nk = 8;
   uint8 Nr = AES256_NR;
